FloodIt Java Application Program
This is a java application representing the game FloodIt, a fun and challenging game where players click on each square to change its color. The goal is to flood the entire board with one color in a limited number of clicks.

To make the game challenging, I implemented a mechanic where the game floods only the squares adjacent to those already flooded, producing a visually appealing waterfall effect. Additionally, if a player clicks on a square that is already the same color as their controlled area, it will not flood.

The game is designed to be proportional in difficulty to the size of the board and number of colors available. If players fail to flood the board within the allotted number of clicks, they will lose. If they succeed, they will win, and the game will notify them accordingly.

For convenience, I have also added a feature that allows players to reset the game and create a new board by simply pressing the 'r' key.